requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
channels>=4
channels-redis
celery>=5
redis
uvicorn
gunicorn
django-cors-headers
django-storages[boto3]
boto3
drf-spectacular
python-dotenv
dj-database-url
pytest
pytest-django
factory-boy
